requirements.txt

django
coverage
python-jose
